Update patch set 1
Patch Set 1:

It looks like that feature is a per-user setting; it stores a full path to the place you designate as containing user dictionaries. We could check in our adt spelling dictionary, but each user would need to go and point to it initially. (This is how things were in Eclipse by the way; we had sdk/eclipse/dictionary.txt and you'd need to point Eclipse to it yourself).

Right now with the adt.xml file checked in, anyone loading the project won't see typos for these issues.

I see +'s and -'s with each approach.

Patch-set: 1