Make view meta model extendable
This changeset makes a couple of methods in the MetaManager protected
rather than private, such that its subclass (such as ViewsMetaManager
in the Android plugin) can override and extend the functionality.

Similarly, the methods which create MetaModels and PaletteItems are
extracted into separate protected methods, such that subclasses can
instantiate subclasses with extra information.

Finally, use the PaletteItem interface rather than the concrete
DefaultPaletteItem in a few places, such that the variation palette
items can be used in place.

Change-Id: Idcb90b9b84cacb29fddff9ee07c4826120f2b636