Allow designer customizations to substitute a palette widget
This changeset (1) adds an interface for the palette panel, (2) makes
client code refer to the interface rather than the concrete
PalettePanel class, and (3) makes the actual palette used be
replaceable by the designer panel. The same thing is done for the
designer (structure and property) panel window.

This will for example allow the Android layout editor to supply a
palette showing render previews (bitmaps of the actual widgets) rather
than icons and text.

Change-Id: Id40bd35b75fb3a06f63d5d1e77530883370bbef4