Update patch set 2
Patch Set 2:

The Android designer will replace not just the tool window content, but the tool window manager as well.
This is done by overriding DesignerEditorPanel#getPaletteWindowManager.
This works in 99% of the cases. The designer code will all call that method, get the right window manager, ask it to produce its content window, etc.
However, there was one exception to this: ToggleEditorModeAction.
That class doesn't have a hook to the corresponding designer, so it just calls the static methods on DesignerToolWindowManager and PaletteToolWindowManager to get the instances. This obviously doesn't work with our overridden methods.
That's why I added the *additional* code in those static getters to consult whether there is a designer customization, and if so, delegate to it to look up what the manager instance should be. In the case of the Android designer, it will return its own window manager instance -- so there isn't a problem with the field created directly above getting combined with this unknown manager.
It's simply a way for the static instance to be able to delegate to another singleton when appropriate.
However, I don't think that the ToggleEditorModeAction is something that a lot of people use. I've uploaded a new patch which rips out this code to support it; now it's *just* the debugger delegation code, which is the really important part.

Patch-set: 2