Update patch set 1
Patch Set 1:

I prefer not to submit it yet. It happen before that the CL is not taken as it is, but modified and then surprise! surprise! merge conflicts. I'll check IDEA's repo on GitHub to see if this CL is already in, if not, I'll ping them. :-)

Patch-set: 1
Reviewer: Gerrit User 1019910 <1019910@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0