Updater: add unit tests.
This adds a few unit tests to the updater.

Also adds a larger UI test to the updater, using it
as a "black box" by creating a patch between 2 mock
data folders and then applying the patch.

The UI test must be invoked using ./gradlew testUI
and since it will use some UI should not be invoked
on a CIS or test server.

SDK Bug: b.android.com/63927

Change-Id: I2630aa9cb838c2b0d83261fd4d8643554a5a345f