Updater: increase delete action timeout to 2 minutes.
On Windows we recently started noticing update fails with a
"Cannot delete file ...\plugins\android\lib\builder-model-0.6.1.jar"
exception. This seems due to the gradle wrapper lingering around for
about 1 minutes after a build. Users start studio, which opens a project,
does a build refresh, then notice an update and apply it... the updater
then runs whilst the gradle wrapper process hasn't finished yet.

Eventually we should wait for the gradle wrapper to finish before
restarting studio to run the update. But on the other hand if we
know a file is being locked and can't be deleted, we can also have
the updater wait a bit longer. The default timeout of 100 ms built
in seems a bit short so this increases it to about 2 minutes for
the delete action.

Change-Id: I204dbf3c53efb89ef269557ef7f473ecc6408bd1