Update patch set 1
Patch Set 1:

By "runtime" you mean "gradle runtime" as in "when gradle runs, e.g. while building code" right? That's what the tests do I think (running gradle on a bunch of different projects with the dev plugin loaded and asserting that the right stuff happens).

Btw the reason I finally got motivated to do this is that there are reports of people running into guava incompatibilities between our plugin and some other plugins. Given that our plugin is using an ancient version of Guava I thought it would be best to update to avoid this.

Patch-set: 1