Update patch set 1
Patch Set 1:

(3 comments)

Can you explain what all the diffs in the Runner class are? The commit message mostly talks about adding unit tests, but it looks like the code is adding support for setting an exit code. Is that just to support unit tests, or is that for something else?

(If it's just for unit test purposes, I've added some code in lint to deal with that (to trap System.exit calls while running the test, to record the exit value, and to assert that it is as expected. See MainTest in tools/base/lint/ - it installs a security manager, traps the system exit calls, and stores it in an exception which it can then check.)

Patch-set: 1