Remove versions from Info.plist
This CL replaces version tags from the Info.plist with a constant
string. This is an attempt to see if the Info.plist file can be
made constant across builds.

This will allow us to use a single signed Mac app (Mac codesigner
only looks at the Contents folder) and not to have to explicitly
sign every build.

Change-Id: If89168930cb071279fb974bdaaa5781ade8b6cdc