log console: Customize list of actions shown in popup menu.
The list of actions shown in the console toolbar is customized
via an extension point, but the list of actions shown in the popup
menu aren't. This CL performs the same customization for the
pop-up menu actions.

This is a fix for https://code.google.com/p/android/issues/detail?id=64180

Change-Id: I0646db950ff391dcc15a34b82009e50f95a52299