Update patch set 1
Patch Set 1:

Our master should reflect IntelliJ 14, not IntelliJ 13.  I'm assuming there was a corresponding fix in IntelliJ 14? Probably best for us to do a new merge to pick up their current trunk instead. I'll handle that.

(If this was an isolated fix in IntelliJ 14 that they then backported to 13 *without making any modifications for 13* then cherrypicking it would be straightforward. But in many cases they'll be picking contents out of files that have changed a lot in the meantime (the 13.x branch is nearly six months old now), so if we put in partial 13 cherrypicks that will conflict with our next 14 merge.

Does that make sense? I'm not sure exactly whether this is the case, but I saw these CLs fly by so I marked them -2 right away to make sure you saw it.

Patch-set: 1