Added a new configuration for running Android plugin tests.
Previously, if you wanted to run the tests for our plugin, you needed to find 
our test folder and kickstart "Run all tests" to automatically create a local 
testing configuration. But then, you needed to stop those tests, because many
would just go on to fail. At this point, you needed to edit the configuration 
and add arguments to the jvm to make sure it started with enough memory to 
handle some of our more expensive tests.

The goal, here, is to have this configuration already built in when someone 
opens a project for the first time. Additionally, I will add a README in a 
different CL to explain the implicit prerequisites our project's tests require 
in order to succeed.

Change-Id: I188e4163379cfe282543b23c0fb41ea44f558217