Adds the ability to right align actions on toolbars.
New attribute on "add-to-group" to indicate the alignment.
Only supported on the top level group or action added to a toolbar (setting it on a subgroup does nothing)
Search is now always anchor=first, alignment=right.  Ideally it should be moved out of code into config.
Implementation is to store an index of the dividing point between left and right and track that when exploded.

Change-Id: I6e2a3c744c8aa13b6fe7cac6dab1da9a18fcdd8a