Update patch set 1
Patch Set 1:

(4 comments)

I've left a couple of comments, but actually I think that this should use a different approach. The action system already has a concept of secondary actions, which is very similar to the right-aligned actions you're adding. Right now they are represented as menu items under a drop-down button. What I would do is add an option to ActionToolbarImpl (or a subclass of ActionToolbar) to show the secondary actions as right-aligned, and set this option for the main toolbar instance (maybe specifically in Android Studio).

Patch-set: 1
Reviewer: Gerrit User 1045657 <1045657@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0