Update patch set 1
Patch Set 1:

Dmitry -- one issue I noticed was that secondary="true" on add-to-group was showing as an error even though it was being parsed.  Should we add it to the schema?

Also I tried to leave current handling for secondaries untouched if the new flag isn't on because there appears to be some experimental stuff going on with auto layout policy.

There should not be a net change to intellij if the new ui setting is left at its default.

Patch-set: 1
Reviewer: Gerrit User 1050224 <1050224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0