Update patch set 1
Patch Set 1:

Yes, but once it is there, someone else might be tempted to use it, and I don't know if everyone would be aware of this. Personally, I'd wait to check with Jetbrains. IF they decide not to, then I'd defer to Xav or Tor to decide whether to take the risk of submitting this and making sure no one in the Android plugin uses this interface via code review.

Patch-set: 1