issue-73938 Add dependency intention doesn't add the dependency to the build.gradle.
https://code.google.com/p/android/issues/detail?id=73938

This CL target IntelliJ platform change, hence, the workflow is a follows:
* let some other team member to reivew the CL
* offer the CL to JetBrains
* wait until the CL is applied to IJ 'master' branch (ensure the change
  is ok)
* abandon current CL
* let the change to be merged into our repo via automatic JB repo sync
  procedure
* apply the change to our IJ release repo fork

So, it's necessary to accomplish the first step now, i.e. review the CL
in order to offer it to JetBrains

We want to be able to sync ide project structure changes with external system
config (gradle in particular). It's convenient to be able to have a handy object-level
representation of the most recent external project state then.

This CL enhances common external system processing in a way to publish an event with
that external project state data to any interested subscriber (it's convenient because at least
for gradle there are different 'refresh project' requesters - like 'refresh project' ide action,
automatic android settings change triggers etc).

Change-Id: I836c4bb47a06618c5652d3c5303a92cd83d4eed4