Create patch set 3
Uploaded patch set 3.

Patch-set: 3
Subject: The updater checks for read/write access to the files and also performs a FileLock check on them. However, on Windows, this is not enough to see if a file can be deleted. If another process holds a lock to a file, neither mechanism will detect it. This change uses a Win API via JNA to detect, and offer the KILL_PROCESS solution, to the user.
Commit: 12d4fc86e3fd26b7139f52db5d4c1c56b550695c