Updater: Adds support for strict patches
For that to happen there are a few other things that needed changing:

The properties file is gone and it's now serialized as part of
the Patch object. The patch object is the one that needs to know
which kind of patch it is so it made more sense to move it in there.

Removed the hack with the global variable for binary patches and
moved it as a property of the patch object. Moved the decision of
which kind of signature should be used to the Patch object, which
is the only one that knows how to do that. This removed the need of
having two subclasses per action type.

Created a PatchSpec object that configures the patch to be created.
This removed a lot of unnecessary arguments and makes the API more
stable if new flags need to be added.

Added the strict mode to the patch. In this mode the following
    things occur:
  . A validation action is created for each file that hasn't changed.
  . The 'solutions' exposed to the user now change if the patch is
    strict.
  . The new solution "NONE" indicates that the patch cannot proceed.
  . For delete actions, if the file to delete is different, we only
    allow to continue if the file is deleted (in no strict mode we
    allow to keep it)
  . For create actions, if there was a file with the same name only
    allow to overwrite it.
  . If a conflict action has no solutions, we don't allow to apply
    the patch.

Note that this doesn't check for unexpected new files, that will be
in a follow up CL.

This CLs also updates the tests so all the tests also run for binary
patches.

The need for a strict mode is because managing installations that
have different binary signatures, or that have files left behind, or
that allow partially updated installations can become an issue.

The overhead of adding validation actions is negligeble in the
overall size of the patch.

Change-Id: Ib633af96332b817b8466d81114384977a78a4feb