Updater: Support for moving files
This change generalizes update actions to be able to take the
data from a different source location. For this to happen update
actions now read from the backup folder to make sure that if there
were other actions applied to the source file, we still have the
original copy. This effectively implements an update-on-move
operation. For a mac example where we moved everything into the
contents folder this brings down patch size from 170MB to 18MB,
which is almost the size it would have had if nothing had moved.

Change-Id: I28d289c3c46badfc332ed061732c90368539cb31