Update patch set 2
Patch Set 2: Code-Review+1 Verified+1

Ok, I just tried this on windows (VM) and it works. However there was another issue where it couldn't delete one of the backup files in the temp folder. This would have happened if a stream is kept open referencing the file. Digging with jvisual vm, I found that there were "ZipFile" objets to it, and all had the "requestToClose" set to true. I also noticed that they have a sleep before deleting a file "waiting" for it to be released. So I added a manual GC call before doing the clean up phase, which fixes the issue after I read this:

http://stackoverflow.com/questions/991489/i-cant-delete-a-file-in-java

PTAL (Runner.java)

Patch-set: 2
Label: Code-Review=+1
Label: Verified=+1