Update patch set 1
Patch Set 1:

I'm not sure if it's worth to do because 'Maven Integration' is listed in the plugin's list when I open Studio settings. I.e. the point is that maven plugin *might* be enabled in particular environment (not sure about default settings though).

So, I'd suggest to wrap that processing in to a helper class which checks if it's safe to proceed.

Btw, it would be good to have stacktrace which shows that CLassNotFoundException problem. May be we should create a ticket for that and attach the trace to it?

Patch-set: 1