Update patch set 1
Patch Set 1:

> We might want to use at least for processing dependencies, i.e.
 > offer auto-completion/inspections for the artifact availability etc

Even if we don't disable Maven central indexing, this functionality is broken. In addition, it would be better to build some incremental index (e.g. make a quick lookup on Maven central and cache that result) than an up-front cache that takes a long time to fill up.

Patch-set: 1