Update patch set 8
Patch Set 8:

(3 comments)

Ok, I reverted the "customizations" in the rc file and in the images. The reason is that this project is supposed to build a "template" which later gets customized to the particular IDE. The customizer has some other issues that need changing (separate CLs) and then the last CLs will be to update our gant and custom images to actually use this template.

Patch-set: 8