Update patch set 1
Patch Set 1:

How come this is in the xpath plugin? Seems like it should be in the Android plugin instead (tools/adt/idea). (When it's in the xpath plugin here, it will only be in the Android Studio fork of IntelliJ, not in IntelliJ itself, so IntelliJ users with the Android plugin won't see these templates).

Patch-set: 1
Reviewer: Gerrit User 1004689 <1004689@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0