Update patch set 1
Patch Set 1:

The main issue with this code is that makes changes to the JetBrains platform unconditionally (e.g. no checks if the project is an Android Gradle project). This change will affect regular Java projects, which can be opened (but not created) in Studio. Since the platform cannot depend on the android plugin, the best way to tackle this is to ask and make a case for an extension point in the platform.

Patch-set: 1