Update patch set 1
Patch Set 1:

I think Tor updated them recently because he is working on a newer version of IDEA. So two different versions of IDEA are going to constantly override each other.

Tor - I had a lot of trouble getting the 14.1 EAP to work properly (it wouldn't finish indexing, the mac app isn't signed, ..). So I'd recommend sticking to the old one for another month or so..

Patch-set: 1