Update patch set 1
Patch Set 1:

It seems to be better on Linux. On a Mac, it was just stuck in the indeterminate progress before it started indexing, maybe it just takes a while longer. I agree that we should move on to the EAP..

Patch-set: 1
Reviewer: Gerrit User 1010055 <1010055@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0