Update patch set 1
Patch Set 1:

Could we have the changes to .idea, .iml and gant files in a separate CL, the image reader change in a separate CL, and the TestStateInfo in a separate CL?

I imagine the first one will forever be a fork from IJ, but we should try and upstream the other 2, so it would be good to have them in independent CLs. Having them separate (with a commit message explaining those changes) makes it easier to get it pulled in by Jetbrains.

Patch-set: 1
Reviewer: Gerrit User 1010055 <1010055@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0