Update patch set 1
Patch Set 1:

Good point about the plugin not being compiled before the dependencies are in! Unfortunately, I just realized that to submit our plugin, I will need to perform some additional work as I was maintaining a forked version of our code and kept changing it according to changes in the extension point (while the original branch was changed to introduce new features and fix bugs) So, as soon the extension point is final (i.e., submitted), I will merge the changes due to work on extension point with feature changes and will submit our plugin.

Patch-set: 1