Introduce RunConfigurationProducerService.
The problem we're trying to solve is that in Android Studio the gradle
test run configurations don't work for neither unit tests nor on-device
tests, and we don't want to support them. After talking to Dmitry, we
concluded that a per-project blacklist of run configuration producers is
a good solution to that. AS will just set the blacklist programatically,
while IJ may provide a UI for editing it.

Change-Id: I03370cb29dfff5e3baa7b4571ecf77147672c264