Update patch set 3
Patch Set 3:

BTW, for any changes in tools/idea, please see if it is possible to upstream that change. Otherwise that part is effectively a fork. This change seems small that we might be ok maintaining that fork, but I'll leave it upto your judgement whether it is something IntelliJ would be interested in.

Patch-set: 3
Reviewer: Gerrit User 1010055 <1010055@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0