Update patch set 1
Patch Set 1:

Internal studio-1.2-dev builds also fail with the same error, but they fail intermittently and I couldn't find a pattern.

I'm also a bit fuzzy on the clean checkout thing, because we don't have the same problem on Jenkins. On the build server though:
- the location of .jps-build-data doesn't depend on the CL
- it gets created on every build if it's not there
- it's also in .gitignore, so I assume it doesn't get wiped out on every build?
I thought this CL is a very cheap thing to try, and we can roll it back tomorrow.

Patch-set: 1