Update patch set 1
Patch Set 1:

The proposal to JetBrains is in a separate CL:
 https://android-review.googlesource.com/149417

That one has *only* the deprecation extension, and there's a separate Android plugin usage which uses the extension to do better checking.

However, it doesn't look like we'll get that in in time for 1.3. And right now the deprecation of ActionBarCompat shows up in nearly every single project created in Android Studio, so I really really want to make it (a) more obvious what the replacement is and (b) have a quickfix for it, so this CL puts part of that functionality into a single CL such that we can submit this without creating an incompatible dependency from the Android plugin to the platform.

I'll revert this and replace with the real extension if it goes in (or if they'd rather we duplicate the deprecation inspection in the Android plugin, I'll do that instead.) So this is a stopgap measure to help with the appcompat deprecation issue.

Patch-set: 1