Add the ability to add more MetaModels to the MetaManager at runtime.
The MetaManager is already available as a service of the Android plugin.
However, at the moment it's only possible to read data from the
MetaManager. This patch allows to add more models. New models have to be
specified in a xml file. For example, this can be used to add more
widgets or layouts to the view palette.

Change-Id: I523ea9619869b9b971dc9d0769cfa85c9f74f553