Introduce DefaultBeforeRunTaskChooser extension point.
This extension point can be used to change the default Make step for
tasks. In Android Studio we will use it to make "Gradle-aware Make" the
default.

The XML representation of run configurations only specifies differences
from the defaults produced in code. This means that the same, empty
"<method />" tag in workspace.xml means "Make" in IJ and "Gradle-aware
Make" in AS.

The extensions ordering mechanism is used to make sure the Android logic
overrides the default logic.

Change-Id: I87f6d5eff5ea331c4f67fc00421caf4b57ae581d