Update patch set 1
Patch Set 1:

(1 comment)

Yeah I considered 1 and 2. For 1) is what happened before, we already had this change "forked" and it got lost in a merge. Forking is not a good idea so I'd like to  push this change to IJ and see what they say.

That took me to 2) I considered adding a system variable and define it in our studio specific file, but that would go as a -D property, and has problems, like it becomes user modifiable and we need to change the way we debug studio to define that property, etc. Adding it to UpdateSettings.State has the same issue as it needs to be initialized somehow. 

If we cannot have an isAndroidStudio, then what we need is a global config object  class that we fork. Bot at build time like the gant file we have, but a java file.

Patch-set: 1