Update patch set 2
Patch Set 2: Code-Review+1

LGTM, I'll leave it to David to +2.

My only comment is that if you are going to submit this in our internal tree, then keep an eye on out for whether IJ picks up this CL. Otherwise you are introducing a new and attractive API (PlatformUtils.isAndroidStudio()) that duplicates an existing API, and someone else might use it in the android plugin and break IJ compatibility.

Patch-set: 2
Reviewer: Gerrit User 1010055 <1010055@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1