Track plugin exceptions
This CL reports exceptions (that get reported on the event log) to GA
if user has consented to sending statistics, or in all internal builds.
Currently, all these exceptions are considered non fatal (since they
don't result in a JVM crash).

For each exception, we only report the exception name itself
(throwable.getClass.getSimpleName) and the location in source where
it was thrown. If the stack trace has "android" in one of its
elements, then the exception is prefixed with "android".

Change-Id: I02da81bf7583e76131b0675688c6059ccc09a9ac