Update patch set 1
Patch Set 1:

This fix looks good to me, but note that this change is in the IntelliJ code base itself, not in the Android plugin. 

The tools/idea repository is, in theory, just a snapshot of the most recent stable version of IntelliJ.   We've made a few tweaks here and there (e.g. changing hardcoded references to IntelliJ over to Android Studio), but in particular, we make sure that we never change the APIs incompatibly; we want to make sure that the Android plugin (in tools/adt/idea) is *ALWAYS* compatible with not just Android Studio but IntelliJ instead.

Therefore, whenever we want to change something in tools/idea, we send the change to JetBrains (via http://youtrack.jetbrains.com/ as a patch); once it's integrated in IntelliJ, if we need it in Studio we apply the patch locally too.

(By the way, this is changing the old layout editor in Android Studio -- as announced at Google I/O 2015, and shown at the end of the tools demo in https://www.youtube.com/watch?v=f7ihSQ44WO0&feature=youtu.be , we're working on a complete rewrite of it.)

Patch-set: 1
Reviewer: Gerrit User 1004689 <1004689@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0